auto-debug-ai/requirements.txt
# AutoGen packages
autogen-core>=0.4.0
autogen-agentchat>=0.4.0
autogen-ext[openai,anthropic,docker,langchain,redis]>=0.4.0

# Core dependencies
python-dotenv>=1.0.0

# Vector database and memory
chromadb>=0.5.0
redis>=5.0.0

# Code analysis and manipulation
ast-grep-py>=0.20.0
unidiff>=0.7.5
black>=24.0.0
pylint>=3.0.0
mypy>=1.10.0
ripgrepy>=2.0.0

# Docker and sandboxing (included in autogen-ext[docker])
docker>=7.0.0

# Web framework (for optional dashboard)
fastapi>=0.111.0
uvicorn>=0.30.0

# Monitoring and observability
prometheus-client>=0.20.0
structlog>=24.0.0
python-json-logger>=2.0.7
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=5.0.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
tenacity>=8.3.0
click>=8.1.0
PyYAML>=6.0.1